Remove MAINTAINERS file
This file is now obsolete now that we have moved to GitLab.  A ticket has been
opened in rtems/docs/rtems-docs#21 to discuss adding a contributors section.

GitLab lists who the 'Code Owners' are on top of every file in the web source
browser this is the canonical way to find out who maintainers are of those files
or directories.

Even with knowing who the maintainers are the best way to figure out who is
looking after what is to open an issue or an MR.  This way we can keep a group
level of maintainership over the code base and have CODEOWNERS enforce approvals
before merging.