open: Fix race condition when called with O_TRUNC
Fix a race condition in which a file opened with O_TRUNC is made
available to other file functions expecting a file descriptor before
the truncation is performed.

This is only possible if the other thread was using an invalid file
descriptor as the open call has yet to finish.

Closes #5109