cpukit/libblock: Close all valid FDs
This is intended to close any valid FDs that were opened, but 0 is also
a valid FD. This change ensures that a FD of 0 is also closed properly.

Coverity CID 1467408 (original)

Coverity CID 1616151 (followup)

off_by_one: Testing whether handle fd is strictly greater than zero is
suspicious. fd leaks when it is zero.