spfatal34/spfatal36: Avoid RTEMS_UNREACHABLE()
The RTEMS_UNREACHABLE() is a hint to the compiler that this code
location is unreachable.  The compiler may use this information to
optimize the code.  If a test case fails and we reach this code
location, then unpredictable things may happen making the test case
debugging more difficult than necessary.