validation: Rework get/set interrupt priorities
For some interrupt controllers like the RISC-V PLIC, a maximum interrupt
priority value of UINT32_MAX is valid. Adjust the validation tests to
take this into account.

In general, all the present interrupt controller API validation tests
have a structural issue. They contain the loop over all interrupt
vectors directly in the action requirement. This is not the right
approach. It is better to use an action requirement which produces a
test run function by defining a test header through the `test-header`
attribute. Then call the test run function for all interrupt vectors.
This approach allows to using the `${.:skip}` feature.

In this commit, this is new approach is applied to the interrupt get/set
priority validation tests.

Close #5578.

Signed-off-by: Sebastian Huber <sebastian.huber@embedded-brains.de>