sparc/include/.../cpuimpl.h: Correct global file scope register use
The register keyword was never more than a suggestion but in more
recent versions of C and C++, it has been reduced to a reserved
keyword with no semantics. GCC has a documented extension for
global register variables. Add the __extension__ keyword to avoid
warnings.

See Issue #5250 for a more detailed discussion including a link
to a discussion on the GCC mailing list.

The use of g6 as a pointer to the per CPU information is documented
in the SPARC chapter of the CPU Supplement Guide.

Closes #5250.