bsps/powerpc/*/include/tm27.h: Correct IRQ installation structure
The gen5200, mvme5500, and psim BSP families had an implementation of
tm27.h which used stub functions with casts as null handlers to fill
in the IRQ structure. The IRQ structure as also missing a field. The
change was to use a set of appropriately signatured stub functions
and correct the structure initialization.

This code likely has produced warnings for a long time but with
GCC 14, these warnings turned into errors.

Closes #5266