spec/aarch64/xen: Reorder includes to fix link
The opto* include must come before the bspopts include or else the flags
from the opto* include get dropped which causes librtemscpu to build
without the function and data sections flag which causes link-time
errors because unused sections cannot be dropped. The mechanism here
should be better understood, but this fixes the build errors.