powerpc/beatnik/include/bsp.h: Disable inport_ and outport_
The inport_[byte|word|long] and outport_[byte|word|long] methods
assume that the port is an address. The Beatnick definitions just
pass through the offset into IO Space. These are broken until they
do that. They do not appear to be used in anything this BSP actually
uses so disabling them works. Code is disabled with a lengthy comment
added if someone needs them to work in the future.