cpukit: Add msg broadcast non-atomicity to doc
rtems_message_queue_broadcast() may, under certain circumstances, copy
the message to tasks which were not waiting on the message queue when
the broadcast started, and may copy the message multiple times to the
same task.

This behaviour is, based on discussion in #4804, something that might
not change in the short term, so expose it in the user documentation.

The wording is copied with modifications from an existing internal
requirements note.

Also correct "none-atomic" -> "non-atomic" in said internal requirements
note.

The changes were auto-generated from rtems-central.

(cherry picked from commit e727b24ea383668203b4cc3a77651cd574b840ea)